Refactor YaraEngine result types to own module
